Fix a bug where types marked in attributes would have their assembly deleted
When an assembly is compiled with mcs and it has an attribute that makes the only use of a type in a different assembly, no reference to that assembly exist.

This leads to the referenced assembly being deleted even though a type in the assembly was marked.

This issue does not happen with roslyn which adds a reference to the other assembly even though it is not strictly necessary.