Fix nunit exceptions thrown during compile
The recent change to compile test cases in parallel led to nunit exceptions thrown from calls such as Assert.Ignore() not propagating back out.  Instead an AggregateException would be thrown and nunit would, in the case of ignore, not fail the test instead of ignore it.

The only tests that do an Assert.Ignore during test case compilation are tests that need to use mcs when running on windows.  We don't run the tests on windows during CI so that is how this slipped by.