Don't mark assembly attributes until the assemblies module is marked
Don't mark top level attrs until the module is marked.  Normally you would never notice this was happening because
the entire assembly would be deleted anyways.  But when you want SweepAssembly to remove as much as possible for an assembly that would normally be Deleted it is helpful if CustomAttributes
in that assembly have not been marked

This PR pulled a change out of a previous PR so that we can continue the discussion from https://github.com/mono/linker/pull/360#discussion_r216666306

Don't mark top level attrs until the module is marked.  Normally you would never notice this was happening because the entire assembly would be deleted anyways.  But when you want SweepAssembly to remove as much as possible for an assembly that would normally be Deleted it is helpful if CustomAttributes in that assembly have not been marked.

For more context, when an assembly is determined to not be needed during the SweepStep we will change the AssemblyAction to Delete and that will cause the assembly to be deleted.  If instead of going down the `SweepStep.RemoveAssembly` code path you changed the AssemblyBack back to Link and called SweepAssembly on that assembly, the only thing that would not be stripped from that assembly are these assembly attributes that I'm trying not to mark.

So worded another way, when an assembly is not used by any code the only thing in that assembly that we DO mark today is CustomAttributes on the AssemblyDefinition and ModuleDefinition.  This doesn't cause any issues today because we use RemoveAssembly instead of SweepAssembly, but it still seems wrong to only mark CustomAttributes.  That is what this PR is fixing.