Fix an edge case with interface sweeping, no instance ctor, and preserve all
When a type has TypePreserve.All it should retain all of it's interfaces.  This was not happening when the type had no instance constructors.  This is a rare border line invalid edge cases, however, we ran into it and PEVerify says the assembly is valid even though the type has no instance constructors.

This PR adds a check during `InitializeType` to ensure this scenario is handled