Fix a bug with mix of copied and linked assemblies
Fix a bug that occurs when an assembly is assigned AssemblyAction.Copy without having ResolveFromAssemblyStep called on it and that assembly references an assembly with AssemblyAction.Link.  When this happens, none of the dependencies of the copied assembly are marked and it's references can be swept away entirely if there are no other assemblies that cause them to be marked.